Make ivy-rich-switch-buffer-path more robust.
Double check that `filename' and `root' are non nil and then use
`string-remove-prefix' to remove `root' from `filename'.

An example of when the old code could blow up would be if you have a
buffer that is visiting file "/tmp/foo", but for whatever reason, the
user (or some other piece of elisp code) has decided to `setq'
`default-directory' to something else which happens to be longer than
the path to the file.

In that case trying to use ivy-rich-switch-buffer-transformer blows up
with "Args out of range".

This patch makes the function more robust to these sorts of weirdness.