Fixed ivy-rich-modify-column issue
The issue is, the columns props are not updated due to the plist-put
call is not set back to `ivy-rich-display-transformers-list`.

If any column is modified, we also need to rebuild the transformer
function. But we don't do backup. We should ONLY backup transformers
before enabling `ivy-rich`, not the ones before modified after when
`ivy-rich` is enabled.